Fixed some issues

+++ src/Kernel.php
@@ -15,10 +15,13 @@ class Kernel extends BaseKernel
     {
         parent::__construct($environment, $debug);
 
-        $filePath = realpath(glob('build/messageBox*.js')[0]);
-        $contents = file_get_contents($filePath);
-        $contents = str_replace('PHP_REPLACE_HOST', $_SERVER['HTTP_HOST'], $contents);
-        file_put_contents($filePath, $contents);
-        copy($filePath, 'build/messageBox.js');
+        $glob = glob('build/messageBox*.js');
+        if ($glob !== false && \array_key_exists(0, $glob)) {
+            $filePath = realpath($glob[0]);
+            $contents = file_get_contents($filePath);
+            $contents = str_replace('PHP_REPLACE_HOST', $_SERVER['HTTP_HOST'], $contents);
+            file_put_contents($filePath, $contents);
+            copy($filePath, 'build/messageBox.js');
+        }
     }
 }
